Apply suggestion

+++ src/impl/CameraEffects/ShakeCameraEffect.cs
@@ -4,11 +4,11 @@ using System;
 public partial class ShakeCameraEffect : TimedCameraEffect
 {
     int Frequency { get; set; }
-    float Amplitude { get; set; }
-    float lastShookTimer = 0;
-    float periodInMs = 0;
-    float previousX = 0;
-    float previousY = 0;
+    double Amplitude { get; set; }
+    double lastShookTimer = 0;
+    double periodInMs = 0;
+    double previousX = 0;
+    double previousY = 0;
     Vector2 lastOffset = new Vector2(0, 0);
 
     public ShakeCameraEffect(float duration, int frequency, float amplitude) : base(duration)
@@ -32,26 +32,26 @@ public partial class ShakeCameraEffect : TimedCameraEffect
         while (lastShookTimer >= periodInMs)
         {
             lastShookTimer -= periodInMs;
-            float intensity = Amplitude * (1.0f - ((Duration - Timer) / Duration));
-            float newX = (float)(rnd.NextDouble() * 2.0f) - 1.0f;  // -1.0 to 1.0
-            float xComponent = intensity * (previousX + ((float)delta * (newX - previousX)));
-            float newY = (float)(rnd.NextDouble() * 2.0f) - 1.0f;  // -1.0 to 1.0
-            float yComponent = intensity * (previousY + ((float)delta * (newY - previousY)));
+            double intensity = Amplitude * (1.0f - ((Duration - Timer) / Duration));
+            double newX = (rnd.NextDouble() * 2.0f) - 1.0f;  // -1.0 to 1.0
+            double xComponent = intensity * (previousX + (delta * (newX - previousX)));
+            double newY = (rnd.NextDouble() * 2.0f) - 1.0f;  // -1.0 to 1.0
+            double yComponent = intensity * (previousY + (delta * (newY - previousY)));
             previousX = newX;
             previousY = newY;
-            Vector2 newOffset = new Vector2(xComponent, yComponent);
+            Vector2 newOffset = new Vector2((float)xComponent, (float)yComponent);
             camera.Offset = camera.Offset - lastOffset + newOffset;
             lastOffset = newOffset;
         }
 
-        lastShookTimer += (float)delta;
+        lastShookTimer += delta;
     }
 
     public override void Start(Camera2D camera)
     {
         base.Start(camera);
         Random rnd = new Random();
-        previousX = (float)(rnd.NextDouble() * 2.0f) - 1.0f;  // -1.0 to 1.0
-        previousY = (float)(rnd.NextDouble() * 2.0f) - 1.0f;  // -1.0 to 1.0
+        previousX = (rnd.NextDouble() * 2.0f) - 1.0f;  // -1.0 to 1.0
+        previousY = (rnd.NextDouble() * 2.0f) - 1.0f;  // -1.0 to 1.0
     }
 }
